+++ R/defaultControl.R
@@ -41,10 +41,11 @@
 #'       \item{\code{leafPhenology [= TRUE]}: Boolean flag to indicate the simulation of leaf phenology for winter-deciduous species.}
 #'       \item{\code{bareSoilEvaporation [= TRUE]}: Boolean flag to indicate the simulation of evaporation from bare soil.}
 #'       \item{\code{unlimitedSoilWater [= FALSE]}: Boolean flag to indicate the simulation of plant transpiration assuming that soil water is always at field capacity.}
-#'       \item{\code{freeDrainage [= TRUE]}: Boolean flag to indicate that the lower boundary condition for water flow is free drainage.}
 #'       \item{\code{unfoldingDD [= 300]}: Degree-days for complete leaf unfolding after budburst has occurred.}
 #'       \item{\code{interceptionMode [= "Gash1995"]}: Infiltration model, either "Gash1995" or "Liu2001".}
 #'       \item{\code{infiltrationMode [= "GreenAmpt1911"]}: Infiltration model, either "GreenAmpt1911" or "Boughton1989".}
+#'       \item{\code{soilDomains [= "single"]}: Either 'single' (for single-domain) or 'dual' (for dual-permeability).}
+#'       \item{\code{freeDrainage [= TRUE]}: Boolean flag to indicate that the lower boundary condition for water flow is free drainage.}
 #'       \item{\code{rhizosphereOverlap [= "total"]}: A string indicating the degree of rhizosphere spatial overlap between plant cohorts:
 #'           \itemize{
 #'             \item{"none" - no overlap (independent water pools).}
@@ -202,9 +203,10 @@ defaultControl<-function(transpirationMode = "Granier") {
     leafPhenology = TRUE,
     bareSoilEvaporation = TRUE,
     unlimitedSoilWater = FALSE,
-    freeDrainage = TRUE,
     interceptionMode = "Gash1995",
     infiltrationMode = "GreenAmpt1911",
+    soilDomains = "single",
+    freeDrainage = TRUE,
     rhizosphereOverlap = "total",
     unfoldingDD = 300,
     verticalLayerSize = 100,