+++ NEWS.md
@@ -23,6 +23,7 @@
 - Saturation excess in soil flows
 - Evaluation of GPP, H and LE
 - Soil thermodynamics revised
+- Single-domain and dual-permeability soil water flows
 
 # Version 3.2.0
 - Correction of a bug arisen in old Rcpp versions